+++ src/refactoring/components/index.ts
@@ -1,1 +1,3 @@
 export * from "./common";
+export * from "./cart";
+export * from "./product";
