+++ src/refactoring/components/product/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductList } from "./ProductList";
