+++ src/refactoring/contexts/CartContext.tsx
@@ -1,0 +1,22 @@
+import { createContext, useContext, ReactNode } from "react";
+
+// hooks
+import { useCart } from "../hooks";
+
+type CartContextType = ReturnType<typeof useCart>;
+
+const CartContext = createContext<CartContextType | null>(null);
+
+export const CartProvider = ({ children }: { children: ReactNode }) => {
+  const value = useCart();
+
+  return <CartContext.Provider value={value}>{children}</CartContext.Provider>;
+};
+
+export const useCartContext = () => {
+  const context = useContext(CartContext);
+  if (!context) {
+    throw new Error("useCartContext must be used within CartProvider");
+  }
+  return context;
+};
